requirements.txt
Django>=5.2
djangorestframework>=3.15
pytz
sqlparse
tzdata
